# openapi.yaml
openapi: "3.0.3"
info:
  title: 股票基金投资论坛 API
  description: |
    股票基金投资论坛后端 RESTful API。
    支持用户认证、内容发布、社交互动、搜索聚合、管理运营等完整功能。
  version: "1.0.0"
  contact:
    name: 贺嘉轩
    url: https://github.com/kkk431/Stock-Fund-Investment-Forum
servers:
  - url: http://localhost:8000
    description: 本地开发环境
  - url: http://api.example.com
    description: 生产环境（待部署）
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 登录后返回的 JWT Token，放入 Authorization 请求头
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          description: 业务状态码
          example: 200
        message:
          type: string
          description: 提示消息
          example: success
        data:
          oneOf:
            - type: object
            - type: array
            - type: "null"
          description: 响应数据
      required: [code, message]
    PaginatedResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          type: object
          properties:
            items:
              type: array
              items:
                type: object
            total:
              type: integer
              example: 100
            page:
              type: integer
              example: 1
            size:
              type: integer
              example: 20
      required: [code, message, data]
...
    CreateCommentRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          maxLength: 2000
        parent_id:
          type: integer
          nullable: true
          description: 父评论ID，一级评论为null
        reply_to_id:
          type: integer
          nullable: true
          description: 回复的目标评论ID

    # ==================== 社交系统 ====================
...
    CollectRequest:
      type: object
      properties:
        folder_name:
          type: string
          description: 收藏文件夹名称

    # ==================== 转发 ====================
    ShareRequest:
      type: object
      required: [share_type]
      properties:
        share_type:
          type: string
          enum: [timeline, message, group]
        comment:
          type: string
          description: 转发附言

    # ==================== 私信 ====================
...
paths:
...
  /api/posts/{id}/comments:
    get:
      tags:
        - 内容系统
      summary: 获取评论列表
      operationId: listComments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponse"
    post:
      tags:
        - 内容系统
      summary: 发表评论
      operationId: createComment
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: 评论成功
  /api/posts/{id}/like:
    post:
      tags:
        - 内容系统
      summary: 点赞/取消点赞
      operationId: toggleLike
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          is_liked:
                            type: boolean
                          like_count:
                            type: integer
  /api/posts/{id}/collect:
    post:
      tags:
        - 内容系统
      summary: 收藏/取消收藏
      operationId: toggleCollect
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          is_collected:
                            type: boolean
                          collect_count:
                            type: integer
  /api/posts/{id}/share:
    post:
      tags:
        - 内容系统
      summary: 转发帖子
      operationId: sharePost
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShareRequest"
      responses:
        "200":
          description: 转发成功
  /api/users/me/collections:
    get:
      tags:
        - 内容系统
      summary: 获取收藏列表
      operationId: listCollections
      security:
        - BearerAuth: []
      parameters:
        - name: folder
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponse"

  # ==================== 社交系统 ====================
...